# openapi/registry.openapi.json
{
  "components": {
    "schemas": {
...
      "AppealRecordInput": {
        "additionalProperties": false,
        "properties": {
          "concern_id": {
            "minLength": 1,
            "title": "Concern Id",
            "type": "string"
          },
          "determination_id": {
            "minLength": 1,
            "title": "Determination Id",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "predecessor_record_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Predecessor Record Id"
          },
          "rationale": {
            "minLength": 10,
            "title": "Rationale",
            "type": "string"
          },
          "record_type": {
            "enum": [
              "appeal",
              "correction"
            ],
            "title": "Record Type",
            "type": "string"
          },
          "requested_by": {
            "minLength": 1,
            "title": "Requested By",
            "type": "string"
          }
        },
        "required": [
          "concern_id",
          "determination_id",
          "record_type",
          "rationale",
          "requested_by"
        ],
        "title": "AppealRecordInput",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Immutable case registry, granular data-use authorization, and blinded scientific concern adjudication, expert calibration, rights-cleared intake, and claim reconstruction.",
    "title": "OpenCritique Registry and Adjudication API",
    "version": "0.5.0-alpha"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/appeals": {
      "post": {
        "operationId": "create_appeal_v1_appeals_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppealRecordInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppealRecordView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Appeal"
      }
    },
    "/v1/artifacts": {
      "post": {
        "operationId": "upload_artifact_v1_artifacts_post",
        "parameters": [
          {
            "in": "query",
            "name": "media_type",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Media Type",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "rights_classification",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/RightsClassification"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Artifact"
      }
    },
    "/v1/artifacts/{sha256}": {
      "get": {
        "operationId": "download_artifact_v1_artifacts__sha256__get",
        "parameters": [
          {
            "in": "path",
            "name": "sha256",
            "required": true,
            "schema": {
              "title": "Sha256",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Artifact"
      }
    },
    "/v1/audit-events": {
      "get": {
        "operationId": "audit_events_v1_audit_events_get",
        "parameters": [
          {
            "in": "query",
            "name": "target_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Target Type"
            }
          },
          {
            "in": "query",
            "name": "target_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Target Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AuditEventView"
                  },
                  "title": "Response Audit Events V1 Audit Events Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Audit Events"
      }
    },
...
    "/v1/cases": {
      "post": {
        "operationId": "register_case_v1_cases_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaseRegistration"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register Case"
      }
    },
    "/v1/cases/{case_id}/versions/{case_version}": {
      "get": {
        "operationId": "get_case_v1_cases__case_id__versions__case_version__get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "case_version",
            "required": true,
            "schema": {
              "title": "Case Version",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseBundle"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Case"
      }
    },
    "/v1/cases/{case_id}/versions/{case_version}/grants": {
      "get": {
        "operationId": "list_grants_v1_cases__case_id__versions__case_version__grants_get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "case_version",
            "required": true,
            "schema": {
              "title": "Case Version",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RightsGrantView"
                  },
                  "title": "Response List Grants V1 Cases  Case Id  Versions  Case Version  Grants Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Grants"
      },
      "post": {
        "operationId": "create_grant_v1_cases__case_id__versions__case_version__grants_post",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "case_version",
            "required": true,
            "schema": {
              "title": "Case Version",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RightsGrantInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RightsGrantView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Grant"
      }
    },
    "/v1/cases/{case_id}/versions/{case_version}/tasks": {
      "get": {
        "operationId": "list_tasks_v1_cases__case_id__versions__case_version__tasks_get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "case_version",
            "required": true,
            "schema": {
              "title": "Case Version",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TaskView"
                  },
                  "title": "Response List Tasks V1 Cases  Case Id  Versions  Case Version  Tasks Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Tasks"
      }
    },
    "/v1/cases/{case_id}/versions/{case_version}/tasks/seed": {
      "post": {
        "operationId": "seed_tasks_v1_cases__case_id__versions__case_version__tasks_seed_post",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "case_version",
            "required": true,
            "schema": {
              "title": "Case Version",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskSeedRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TaskView"
                  },
                  "title": "Response Seed Tasks V1 Cases  Case Id  Versions  Case Version  Tasks Seed Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Seed Tasks"
      }
    },
...
    "/v1/concerns/{concern_id}/appeals": {
      "get": {
        "operationId": "list_appeals_v1_concerns__concern_id__appeals_get",
        "parameters": [
          {
            "in": "path",
            "name": "concern_id",
            "required": true,
            "schema": {
              "title": "Concern Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AppealRecordView"
                  },
                  "title": "Response List Appeals V1 Concerns  Concern Id  Appeals Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Appeals"
      }
    },
    "/v1/concerns/{concern_id}/determination": {
      "get": {
        "operationId": "get_determination_v1_concerns__concern_id__determination_get",
        "parameters": [
          {
            "in": "path",
            "name": "concern_id",
            "required": true,
            "schema": {
              "title": "Concern Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeterminationView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Determination"
      }
    },
...
  }
}